backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
qdrant-client==1.8.0
sentence-transformers==2.2.2
numpy==1.24.3
aiofiles==23.2.1
python-markdown==3.5.1
beautifulsoup4==4.12.2
tiktoken==0.5.2
torch>=2.0.0
transformers>=4.35.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3
websockets==12.0